rename sprite and add min sprite
